requirements.txt
streamlit
phidata
google-generativeai
openai
duckduckgo-search
pytube
yt-dlp
bs4